% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chelsa_Download_functions.R
\name{Chelsa.Clim.download}
\alias{Chelsa.Clim.download}
\title{Function for downloading the Chelsa Climate dataset}
\usage{
Chelsa.Clim.download(
  save.location = "./",
  parameter = c("prec", "temp", "tmax", "tmin", "bio"),
  bio.var = c(1:19),
  month.var = c(1:12),
  version.var = c("1.2"),
  clipping = FALSE,
  clip.shapefile = NULL,
  clip.extent = c(-180, 180, -90, 90),
  buffer = 0,
  convert.files.to.asc = FALSE,
  stacking.data = FALSE,
  combine.raw.zip = FALSE,
  delete.raw.data = FALSE,
  save.bib.file = TRUE
)
}
\arguments{
\item{save.location}{string. Input where the datasets should be saved. \cr Default: Working Directory.}

\item{parameter}{string (vector). Input of parameters which should be downloaded. \cr Default: \code{c("prec", "temp", "tmax", "tmin", "bio")}}

\item{bio.var}{integer (vector). Input which monthly data should be downloaded. Only applicable to BIOCLIM variables. For further information see: \url{http://chelsa-climate.org/bioclim/}. \cr Default: \code{c(1:19)}}

\item{month.var}{integer (vector). Input which monthly data should be downloaded. Only applicable to precipitation and temperature (average, maximum, minimum). \cr Default: \code{c(1:12)}}

\item{version.var}{string (vector). Input which version of the data set should be downloaded. Multiple selection is possible. \cr Default:  \code{c("1.1", "1.2")}}

\item{clipping}{logical. Input whether the downloaded data should be clipped.\cr If \code{FALSE}: clip.shapefile, buffer, clip.extent will be ignored. \cr Default: \code{FALSE}}

\item{clip.shapefile}{string. Input which shapefile should be used for clipping. \cr Default: \code{NULL}}

\item{clip.extent}{numeric (vector). Input vector with four numeric values. This is following the input order c("xleft", "xright", "ybottom", "ytop"). \cr Default: \code{c(-180, 180, -90, 90)}}

\item{buffer}{numeric. Input of decimal degrees of buffer around the shapefile and/or extent. \cr Default: \code{0}}

\item{convert.files.to.asc}{logical. Input whether files should be converted into the ASCII format.\cr If \code{TRUE}: a new subdirectory is created and the rawdata is saved there. If \code{clipping} is \code{TRUE}: the clipped raster files are also saved as ASCII grids. \cr  Default: \code{FALSE}}

\item{stacking.data}{logical. Input whether the downloaded data should be stacked as a netCDF-rasterstack. \cr Default: \code{FALSE}}

\item{combine.raw.zip}{logical. Should the downloaded raw-data be "zipped". \cr Default: \code{FALSE}}

\item{delete.raw.data}{logical. Should the downloaded raw-data be deleted.\cr If \code{combine.raw.zip} is \code{TRUE}: raw-data is still available in the zipped file. \cr Default: \code{FALSE}}

\item{save.bib.file}{logical. Whether a bibTex-citation file of the data set should be provided in the Working directory. \cr Default: \code{TRUE}}
}
\value{
Chelsa climate data sets for the period of 1979 - 2013
}
\description{
This function supports a download of the Chelsa Climate dataset. This includes precipitation (mm), temperature (average, maximum, minimum; °C), and bioclimatic parameters. For convenience also a clipping-, conversion to ascii-grid,and stacking-tool is included.\cr An output of a .bib-file of the literature is also optional.\cr For a clear working environment, directories will be created automatically. Also options to "zip" and/or delete the RAW-files are included.
}
\examples{
## NOT RUN
## Bioclim
# Chelsa.Clim.download(parameter = "bio", bio.var = c(1,12))
## Precipitation
# Chelsa.Clim.download(parameter = "prec", month.var = c(1,12))
## NOT RUN

}
\author{
Helge Jentsch
}
