% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chelsa_Download_functions.R
\name{Chelsa.lgm.download}
\alias{Chelsa.lgm.download}
\title{Function for downloading CHELSA Last Glacial Maximum datasets}
\usage{
Chelsa.lgm.download(
  save.location = "./",
  parameter = c("prec", "temp", "tmax", "tmin", "bio"),
  bio.var = c(1:19),
  month.var = c(1:12),
  model.var = c("CCSM4", "MRI-CGCM3", "CNRM-CM5", "FGOALS-g2", "IPSL-CM5A-LR",
    "MIROC-ESM", "MPI-ESM-P"),
  download.dem = FALSE,
  clipping = FALSE,
  clip.shapefile = NULL,
  clip.extent = c(-180, 180, -90, 90),
  buffer = 0,
  convert.files.to.asc = FALSE,
  stacking.data = FALSE,
  combine.raw.zip = FALSE,
  delete.raw.data = FALSE,
  save.bib.file = TRUE
)
}
\arguments{
\item{save.location}{string. Input where the datasets should be saved. \cr Default: Working Directory.}

\item{parameter}{string (vector). Input of parameters which should be downloaded. \cr Default: \code{c("prec", "temp", "tmax", "tmin", "bio")}}

\item{bio.var}{integer (vector). Input which bioclim data should be downloaded. Only applicable to BIOCLIM variables. For further information see: \url{https://www.chelsa-climate.org/bioclim/}. \cr Default: \code{c(1:19)}}

\item{month.var}{integer (vector). Input which monthly data should be downloaded. Only applicable to Precipitation and Temperature (average, maximum, minimum). \cr Default: \code{c(1:12)}}

\item{model.var}{string (vector). Input which future model dataset should be downloaded. For more information see: \url{https://www.chelsa-climate.org/last-glacial-maximum-climate/}. For some of the datasets not all downloads are available. For the ones that are not supported the data will not be downloaded and a warning will be prompted. For an overview please try "warnings()" after execution. \cr Default: \code{c("CCSM4", "MRI-CGCM3", "CNRM-CM5", } \cr \code{"FGOALS-g2", "IPSL-CM5A-LR", "MIROC-ESM", "MPI-ESM-P")}}

\item{download.dem}{logical. Input whether a LGM digital elevation model should be downloaded. \cr Default: \code{FALSE}}

\item{clipping}{logical. Input whether the downloaded data should be clipped.\cr If \code{FALSE}; clip.shapefile, buffer, clip.extent will be ignored. \cr Default: \code{FALSE}}

\item{clip.shapefile}{string. Input which shapefile should be used for clipping. \cr Default: \code{NULL}}

\item{clip.extent}{numeric (vector). Input vector with four numeric values. This is following the input order c("xleft", "xright", "ybottom", "ytop").\cr Default: \code{c(-180, 180, -90, 90)}}

\item{buffer}{numeric. Input of decimal degrees of buffer around the shapefile and/or extent. \cr Default: \code{0}}

\item{convert.files.to.asc}{logical. Input whether files should be converted into the ASCII format. If \code{TRUE}: a new subdirectory is created and the rawdata is saved there. \cr If \code{clipping} is \code{TRUE}: the clipped raster files are also saved as ASCII grids. \cr  Default: \code{FALSE}}

\item{stacking.data}{logical. Input whether the downloaded data should be stacked as a netCDF-rasterstack. \cr Default: \code{FALSE}}

\item{combine.raw.zip}{logical. Should the downloaded raw-data be "zipped". \cr Default: \code{FALSE}}

\item{delete.raw.data}{logical. Should the downloaded raw-data be deleted. If the "combine.raw.zip"-option is \code{TRUE}, raw-data is still available in the zipped file.\cr Default: \code{FALSE}}

\item{save.bib.file}{logical. Whether a BibTex-citation file of the CHELSA dataset should be provided in the Working directory. \cr Default: \code{TRUE}}
}
\value{
Downscaled global climatological data from the last glacial maximum.
}
\description{
This function supports a download of the CHELSA Last Glacial Maximum Climate datasets (21.000 BP). This includes monthly precipitation sums in mm, monthly temperature (average, maximum, minimum) in degree Celsius, annual characteristics (19 bioclimatic parameters), and a global digital elevation model. For further information, please regard \url{https://www.chelsa-climate.org/last-glacial-maximum-climate/}.\cr To allow pre-processing, clipping and buffering, conversion to ASCII-grids and stacking options are included.\cr Optional an output of a .bib-file of the cited literature can be retrieved.\cr For user convenience, saving directories will be created automatically. Also options to "zip" and/or delete the RAW-files are included.
}
\details{
"The CHELSA LGM data is based on a implementation of the CHELSA algorithm on PMIP3 data." (CHELSA Climate 2020: \url{https://www.chelsa-climate.org/last-glacial-maximum-climate/})
}
\note{
Please note that the downloaded data for temperature and the therefore also the first eleven bioclim-variables are processed to °C with one significant decimal without offset and factor. Processing and conversion to other file-formats on a global dataset may take some time.\cr For some of the datasets not all models are available. For the ones that are not supported the data will not be downloaded and a warning will be prompted. See parameter \code{model.var} for more information or check the website of CHELSA Climate (\url{https://www.chelsa-climate.org/last-glacial-maximum-climate/}).
}
\examples{
\dontrun{
# Bioclim
Chelsa.lgm.download(parameter = "bio",
                    bio.var = c(1,19),
                    model.var = "MPI-ESM-P")
# Precipitation
Chelsa.lgm.download(parameter = "prec",
                    month.var = c(1,12),
                    model.var = "MPI-ESM-P")
}


}
\references{
D. N. Karger, O. Conrad, J. Böhner , et al. "Climatologies at high resolution for the earth's land surface areas". In: _Scientific Data_ 4.1 (Sep. 2017). DOI: 10.1038/sdata.2017.122. <URL: <https://doi.org/10.1038/sdata.2017.122>>.

D. N. Karger, O. Conrad, J. Böhner , et al. _Data from: Climatologies at high resolution for the earth's land surface areas_. En. 2018. DOI: 10.5061/DRYAD.KD1D4. <URL: <https://datadryad.org/stash/dataset/doi:10.5061/dryad.kd1d4>>.
}
\author{
Helge Jentsch
}
