% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{clipping.tif}
\alias{clipping.tif}
\title{Clipping .tif-raster files in one specified directory}
\usage{
clipping.tif(
  clip.save.location = "./",
  clip.shapefile = NULL,
  clip.extent = c(-180, 180, -90, 90),
  buffer = 0,
  convert.files.to.asc = FALSE,
  time.stamp.var = str_replace_all(str_replace_all(paste0(Sys.time()), pattern = ":",
    replacement = "-"), pattern = " ", replacement = "_")
)
}
\arguments{
\item{clip.save.location}{string (directory path). The directory where .tif-raster files are saved. \cr Default: \code{"./"} (Working directory)}

\item{clip.shapefile}{string (file path to a ESRI shapefile with the file extension ".shp"). \cr Extent of this shapefile is used to clip the .tif-raster files if no buffer is specified. \cr Default: \code{NULL}}

\item{clip.extent}{numeric (vector). Input vector with four numeric values. This is following the input order c("xleft", "xright", "ybottom", "ytop").\cr It is used if no shapefile input is specified. If also left unspecified, the maximum extent (of the raster file) is used.\cr Default: \code{c(-180, 180, -90, 90)}}

\item{buffer}{numeric. Input of decimal degrees of buffer around the shapefile and/or extent. \cr Default: \code{0}}

\item{convert.files.to.asc}{logical. Input whether the clipped output should be converted into ASCII-Grids. \cr Default: \code{FALSE}}

\item{time.stamp.var}{string. Timestamp to create unique directories for multiple run outputs. \cr Default: \code{stringr::str_replace_all(stringr::str_replace_all(paste0(} \cr \code{Sys.time()),pattern = ":",replacement = "-"))}}
}
\value{
This function returns, depending on the parameter \code{convert.files.to.asc} whether it is ASCII- or GEOTIFF-format, clipped raster files to a new directory. This directory is dynamically created.
}
\description{
This function clips all .tif-raster files in a specified directory (clip.save.location) to an extent given by a shapefile or user input with the possibility of adding a buffer. Furthermore an option for conversion to ASCII format is given.
}
\examples{
#clipping.tif(clip.save.location = system.file("pictures/", package = "rgdal"))

}
\author{
Helge Jentsch
}
