% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{combine.raw.in.zip}
\alias{combine.raw.in.zip}
\title{Combines all .tif-raster files into a .zip-file}
\usage{
combine.raw.in.zip(
  save.location = "./",
  zip.name = "RAWDATA",
  unique.name = TRUE,
 
    time.stamp.var = stringr::str_replace_all(stringr::str_replace_all(paste0(Sys.time()),
    pattern = ":", replacement = "-"), pattern = " ", replacement = "_")
)
}
\arguments{
\item{save.location}{string (directory path). The directory where .tif-raster files are saved and the created/updated .zip file should be saved. \cr Default: \code{"./"} (Working directory)}

\item{zip.name}{string. Input how the .zip-file should be named. \cr Default: \code{"RAWDATA"}}

\item{unique.name}{logical. Should the .zip-file be named uniquely? If TRUE the current system time is added as a timestamp to create unique directories for multiple run outputs. \cr Default: \code{TRUE}}

\item{time.stamp.var}{string. Input of current system time or, if called within another function the initial time of the execution. \cr Default: \code{stringr::str_replace_all(stringr::str_replace_all(paste0(} \cr \code{Sys.time()),pattern = ":",replacement = "-"))}}
}
\description{
Combines all .tif-raster files of one directory into a .zip-file with a custom name.
}
\author{
Helge Jentsch
}
