% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{convert.to.asc}
\alias{convert.to.asc}
\title{Converting .tif-raster files in one specified directory into ASCII-Grids}
\usage{
convert.to.asc(
  save.location = "./",
  time.stamp.var = str_replace_all(str_replace_all(paste0(Sys.time()), pattern = ":",
    replacement = "-"), pattern = " ", replacement = "_")
)
}
\arguments{
\item{save.location}{string (directory path). The directory where .tif-raster files are saved. \cr Default: \code{"./"} (Working directory)}

\item{time.stamp.var}{string. Timestamp to create unique directories for multiple run outputs. \cr Default: \code{stringr::str_replace_all(stringr::str_replace_all(paste0(} \cr \code{Sys.time()),pattern = ":",replacement = "-"))}}
}
\value{
This function returns ASCII-format raster files to a new directory. This directory is dynamically created.
}
\description{
This function converts and saves all .tif-raster files to ASCII-file format to a specified location and creates a new folder named "ASCII_files" with a unique timestamp of the current system time.
}
\examples{
convert.to.asc(save.location = system.file("pictures/", package = "rgdal"))
}
\author{
Helge Jentsch
}
