% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{getDownloadSize}
\alias{getDownloadSize}
\title{Get Download Size}
\usage{
getDownloadSize(URLVector)
}
\arguments{
\item{URLVector}{Character vector. Multiple vectors of valid URLs.}
}
\value{
Download size as double numeric value
}
\description{
Helper function that returns the download size of a vector of URLs
}
\author{
Helge Jentsch
}
