% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{process.raster.offset}
\alias{process.raster.offset}
\title{Preprocessing data to undo offsetting values}
\usage{
process.raster.offset(raster.layer = NULL, offset = -273.15)
}
\arguments{
\item{raster.layer}{SpatRaster to be processed.}

\item{offset}{Numerical value to be offsetted. Default: -273.15.}
}
\value{
SpatRaster
}
\description{
Takes input SpatRaster, deletes offset, and returns the SpatRaster.
}
\author{
Helge Jentsch
}
