% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{save.citation}
\alias{save.citation}
\title{Save the citation of the downloaded dataset}
\usage{
save.citation(
  save.location = "./",
  dataSetName = c("Chelsa", "WorldClim1.4", "WorldClim2.1")
)
}
\arguments{
\item{save.location}{string (directory path). Where the BibTex-file will be saved. \cr Default: \code{"./"} (Working Directory)}

\item{dataSetName}{string (vector). Specifies which dataset was downloaded or which citation should be saved. \cr Default: \code{c("Chelsa1.2", "WorldClim1.4", "WorldClim2.1")} (all available datasets)}
}
\value{
BIBtex-file with biliography of the downloaded dataset
}
\description{
Saves the citation of the downloaded dataset into a BibTex-file in the Working Directory.
}
\note{
DISCLAIMER: No warranty or liability! The citations are provided without any warranty of any kind whatsoever, either expressed or implied, including warranties of merchantability and fitness for a particular purpose. The author should not be responsible for any incomplete citation of datasets or climate data products downloaded through this package.
}
\author{
Helge Jentsch
}
