% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{stacking.downloaded.data}
\alias{stacking.downloaded.data}
\title{Stacking .tif-raster files in one specified directory}
\usage{
stacking.downloaded.data(
  stack.save.location = "./",
  stack.clipped = FALSE,
  variable.numbers = c(1:length(list.files("./", pattern = ".tif"))),
 
    time.stamp.var = stringr::str_replace_all(stringr::str_replace_all(paste0(Sys.time()),
    pattern = ":", replacement = "-"), pattern = " ", replacement = "_")
)
}
\arguments{
\item{stack.save.location}{string (directory path). The directory where .tif-raster files are saved. Rasterfiles must be in a Geographic Coordinate System (in Arc-Degrees) \cr Default: \code{"./"} (Working directory)}

\item{stack.clipped}{logical. Input whether clipped data should be stacked and saved as netCDF as well. \cr Default: \code{FALSE}}

\item{variable.numbers}{numeric (vector). Input how the stack variables should be called. By default just a sequence from 1 to the number of ".tif"-files are used.\cr Default: \code{c(1:length(list.files("./", pattern = ".tif")))}}

\item{time.stamp.var}{string. Timestamp to create unique directories for multiple run outputs.\cr Default: \code{stringr::str_replace_all(stringr::str_replace_all(paste0(} \cr \code{Sys.time()),pattern = ":",replacement = "-"))}}
}
\description{
This function stacks all .tif-raster files in a specified directory and saves the Raster*stack File as netCDF-file in the specified directory.
}
\author{
Helge Jentsch
}
