% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldClim_download_functions.R
\name{WorldClim.CMIP_5.download}
\alias{WorldClim.CMIP_5.download}
\title{Function for downloading the WorldClim v1.4 CMIP5 Future Climate dataset}
\usage{
WorldClim.CMIP_5.download(
  save.location = "./",
  parameter = c("prec", "tmax", "tmin", "bio"),
  bio.var = c(1:19),
  month.var = c(1:12),
  resolution = c("10m", "5m", "2.5m", "30s"),
  model.var = c("ACCESS1-0", "BCC-CSM1-1", "CCSM4", "CESM1-CAM5-1-FV2", "CNRM-CM5",
    "GFDL-CM3", "GFDL-ESM2G", "GISS-E2-R", "HadGEM2-AO", "HadGEM2-CC", "HadGEM2-ES",
    "INMCM4", "IPSL-CM5A-LR", "MIROC-ESM-CHEM", "MIROC-ESM", "MIROC5", "MPI-ESM-LR",
    "MRI-CGCM3", "NorESM1-M"),
  emission.scenario.var = c("rcp26", "rcp45", "rcp60", "rcp85"),
  time.interval.var = c("2050", "2070"),
  clipping = FALSE,
  clip.shapefile = NULL,
  clip.extent = c(-180, 180, -90, 90),
  buffer = 0,
  convert.files.to.asc = FALSE,
  stacking.data = FALSE,
  keep.raw.zip = FALSE,
  delete.raw.data = FALSE,
  save.bib.file = TRUE
)
}
\arguments{
\item{save.location}{string. Input where the datasets should be saved. \cr Default: Working Directory.}

\item{parameter}{string (vector). Input of parameters which should be downloaded. \cr Default: \code{c("prec", "tmax", "tmin", "bio")}}

\item{bio.var}{integer (vector). Input which monthly data should be downloaded. Only applicable to BIOCLIM variables. For further information see: \url{https://www.worldclim.org/data/bioclim.html}. \cr Default: \code{c(1:19)}}

\item{month.var}{integer (vector). Input which monthly data should be downloaded. Only applicable to precipitation and temperature (maximum, minimum). \cr Default: \code{c(1:12)}}

\item{resolution}{string (vector). Ranging from a 10 arc-minute resolution over 5 and 2.5 arc-minute to 30 arc-second resolution.\cr Default: \code{c("10m", "5m", "2.5m", "30s")}}

\item{model.var}{string (vector). Model used to calculate the dataset.\cr Default: \code{c("ACCESS1-0", "BCC-CSM1-1", "CCSM4", "CESM1-CAM5-1-FV2", "CNRM-CM5", }\cr \code{"GFDL-CM3", "GFDL-ESM2G", "GISS-E2-R", "HadGEM2-AO", "HadGEM2-CC", "HadGEM2-ES",} \cr \code{"INMCM4", "IPSL-CM5A-LR", "MIROC-ESM-CHEM", "MIROC-ESM", "MIROC5",} \cr \code{"MPI-ESM-LR","MRI-CGCM3", "NorESM1-M")}}

\item{emission.scenario.var}{string (vector). Input which emission scenario dataset should be downloaded. Provided are the representative concentration pathways (RCP) 2.6, 4.5, 6.0, and 8.5. \cr Default: \code{c("rcp26", "rcp45", "rcp60", "rcp85")}}

\item{time.interval.var}{string (vector). Time interval for which the dataset is calculated. The given intervals are 2040-2060 (represented by "2050"), and 2060 to 2080 (represented by "2070"). \cr Default: \code{c("2050", "2070")}}

\item{clipping}{logical. Input whether the downloaded data should be clipped. \cr If \code{FALSE}; clip.shapefile, buffer, clip.extent will be ignored. \cr Default: \code{FALSE}}

\item{clip.shapefile}{string. Input which shapefile should be used for clipping. \cr Default: \code{NULL}}

\item{clip.extent}{numeric (vector). Input vector with four numeric values. This is following the input order c("xleft", "xright", "ybottom", "ytop").\cr Default: \code{c(-180, 180, -90, 90)}}

\item{buffer}{numeric. Input of decimal degrees of buffer around the shapefile and/or extent. \cr Default: \code{0}}

\item{convert.files.to.asc}{logical. Input whether files should be converted into the ASCII format.\cr If \code{TRUE} a new subdirectory is created and the rawdata is saved there.\cr If \code{clipping} is \code{TRUE} the clipped raster files are also saved as ASCII grids.\cr Default: \code{FALSE}}

\item{stacking.data}{logical. Input whether the downloaded data should be stacked as a netCDF-rasterstack.\cr Default: \code{FALSE}}

\item{keep.raw.zip}{logical. Should the downloaded raw-data be "zipped".\cr Default: \code{FALSE}}

\item{delete.raw.data}{logical. Should the downloaded raw-data be deleted.\cr If the \code{combine.raw.zip}-option is \code{TRUE}, raw-data is still available in the zipped file.\cr Default: \code{FALSE}}

\item{save.bib.file}{logical. Whether a bibTex-citation file of the dataset should be provided in the Working directory.\cr Default: \code{TRUE}}
}
\value{
WorldClim 1.4 CMIP5 Future climate data sets for the periods of 2041-2060 and/or 2061-2080.
}
\description{
This function supports a download of the WorldClim v1.4 CMIP5 Future Climate dataset. This includes precipitation (mm), temperature (maximum, minimum; °C), and bioclimatic parameters. For convenience also a clipping-, conversion to ascii-grid, and stacking-tool is included. \cr An output of a .bib-file of the literature is also optional. \cr For a clear working environment, directories will be created automatically. Also options to "zip" and/or delete the RAW-files are included.
}
\note{
Please note, that this dataset is regarded to as "outdated" by the WorldClim creators. The download of the current dataset "CMIP6" is also provided by this package with the \code{\link{WorldClim.CMIP_6.download}} function.
}
\author{
Helge Jentsch
}
