% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chelsa_new_download_functions.R
\name{Chelsa.Clim.download}
\alias{Chelsa.Clim.download}
\title{Function for downloading the CHELSA climate dataset (1979-2013)}
\usage{
Chelsa.Clim.download(
  save.location = "./",
  parameter = c("prec", "temp", "tmax", "tmin", "bio"),
  bio.var = c(1:19),
  month.var = c(1:12),
  version.var = c("1.2", "2.1"),
  clipping = FALSE,
  clip.shapefile = NULL,
  clip.extent = c(-180, 180, -90, 90),
  buffer = 0,
  convert.files.to.asc = FALSE,
  stacking.data = FALSE,
  combine.raw.zip = FALSE,
  delete.raw.data = FALSE,
  save.bib.file = TRUE,
  save.download.table = TRUE
)
}
\arguments{
\item{save.location}{string. Input where the datasets will be saved. \cr Default: Working Directory.}

\item{parameter}{string (vector). Input of parameters which should be downloaded. \cr Default: \code{c("prec", "temp", "tmax", "tmin", "bio")}}

\item{bio.var}{integer (vector). Input which monthly data should be downloaded. Only applicable to BIOCLIM variables. For further information see: \url{https://www.chelsa-climate.org/bioclim/}. \cr Default: \code{c(1:19)}}

\item{month.var}{integer (vector). Input which monthly data should be downloaded. Only applicable to precipitation and temperature (average, maximum, minimum). \cr Default: \code{c(1:12)}}

\item{version.var}{string (vector). Input which version of the dataset should be downloaded. Multiple selection is possible. \cr Default:  \code{c("1.2")}}

\item{clipping}{logical. Input whether the downloaded data should be clipped.\cr If \code{FALSE}: clip.shapefile, buffer, clip.extent will be ignored. \cr Default: \code{FALSE}}

\item{clip.shapefile}{string. Input which shapefile should be used for clipping. \cr Default: \code{NULL}}

\item{clip.extent}{numeric (vector). Input vector with four numeric values. This is following the input order c("xleft", "xright", "ybottom", "ytop"). \cr Default: \code{c(-180, 180, -90, 90)}}

\item{buffer}{numeric. Input of decimal degrees of buffer around the shapefile and/or extent. \cr Default: \code{0}}

\item{convert.files.to.asc}{logical. Input whether files should be converted into the ASCII format.\cr If \code{TRUE}: a new subdirectory is created and the rawdata is saved there. If \code{clipping} is \code{TRUE}: the clipped raster files are also saved as ASCII grids. \cr  Default: \code{FALSE}}

\item{stacking.data}{logical. Input whether the downloaded data should be stacked as a netCDF-rasterstack. \cr Default: \code{FALSE}}

\item{combine.raw.zip}{logical. Should the downloaded raw-data be "zipped". \cr Default: \code{FALSE}}

\item{delete.raw.data}{logical. Should the downloaded raw-data be deleted.\cr If \code{combine.raw.zip} is \code{TRUE}: raw-data is still available in the zipped file. \cr Default: \code{FALSE}}

\item{save.bib.file}{logical. Whether a BibTex-citation file of the dataset should be provided in the Working directory. \cr Default: \code{TRUE}}

\item{save.download.table}{logical. Whether a table containing the download infos should be saved. \cr Default: \code{TRUE}}
}
\value{
CHELSA climate datasets for the period of 1979 - 2013
}
\description{
This function supports the download, pre-processing and management of CHELSA climate data comprising of monthly precipitation sums in mm, monthly temperature (average, minimum, maximum) in degrees Celsius, and annual chracteristics (19 bioclimatic variables). The spatial resolution of the downloaded data is 30 arc-seconds.\cr To allow pre-processing, clipping and buffering, conversion to ASCII-grids and stacking options are included.\cr Optional an output of a .bib-file of the cited literature can be retrieved.\cr For user convenience, saving directories will be created automatically. Also options to "zip" and/or delete the RAW-files are included.
}
\note{
Please note that the downloaded data for temperature and the first eleven bioclim-variables are processed to °C with one significant decimal without offset and factor. Processing and conversion to other file-formats on a global dataset may take some time.
}
\examples{
\dontrun{
# Bioclim
Chelsa.Clim.download(parameter = "bio", bio.var = c(1,19))
# Precipitation
Chelsa.Clim.download(parameter = "prec", month.var = c(1,12))
}

}
\references{
D. N. Karger, O. Conrad, J. Böhner , et al. _Climatologies at high resolution for the earth's land surface areas_. In: _Scientific Data_ 4.1 (Sep. 2017). DOI: 10.1038/sdata.2017.122. <URL: https://doi.org/10.1038/sdata.2017.122>.

D. N. Karger, O. Conrad, J. Böhner , et al. _Data from: Climatologies at high resolution for the earth's land surface areas_. En. 2018. DOI: 10.5061/DRYAD.KD1D4. <URL: https://datadryad.org/stash/dataset/doi:10.5061/dryad.kd1d4>.

D. N. Karger, O. Conrad, J. Böhner , et al. _Climatologies at high resolution for the earth’s land surface areas_ EnviDat. (2021) DOI: 10.16904/envidat.228. <URL: https://www.doi.org/10.16904/envidat.228>.
}
\author{
Helge Jentsch
}
