% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{process.raster.int.doub}
\alias{process.raster.int.doub}
\title{Preprocessing data to get real values}
\usage{
process.raster.int.doub(raster.layer = NULL)
}
\arguments{
\item{raster.layer}{RasterLayer to be processed}
}
\value{
RasterLayer
}
\description{
Takes input RasterLayer, crops it, processes the integer values into double values, mosaiks the cropped and processed data, and returns the mosaiked Rasterfile
}
\author{
Helge Jentsch
}
