% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{process.raster.int.doub}
\alias{process.raster.int.doub}
\title{Preprocessing data to get real values}
\usage{
process.raster.int.doub(raster.layer = NULL)
}
\arguments{
\item{raster.layer}{SpatRaster to be processed}
}
\value{
SpatRaster
}
\description{
Takes input SpatRaster, processes the integer values into double values, and returns the SpatRaster.
}
\author{
Helge Jentsch
}
