% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldClim_download_functions.R
\name{WorldClim.HistClim.download}
\alias{WorldClim.HistClim.download}
\title{Function for downloading the WorldClim historical climate dataset}
\usage{
WorldClim.HistClim.download(
  save.location = "./",
  parameter = c("prec", "temp", "tmax", "tmin", "srad", "wind", "vapr", "bio", "elev"),
  bio.var = c(1:19),
  month.var = c(1:12),
  resolution = c("10min", "5min", "2.5min", "30s"),
  version.var = c("1.4", "2.1"),
  clipping = FALSE,
  clip.shapefile = NULL,
  clip.extent = c(-180, 180, -90, 90),
  buffer = 0,
  convert.files.to.asc = FALSE,
  stacking.data = FALSE,
  keep.raw.zip = FALSE,
  delete.raw.data = FALSE,
  save.bib.file = TRUE
)
}
\arguments{
\item{save.location}{string. Input where the datasets should be saved. \cr Default: Working Directory.}

\item{parameter}{string (vector). Input of parameters which should be downloaded. \cr Default: \code{c("prec", "temp", "tmax","tmin", "srad", "wind", "vapr", "bio", "elev").}}

\item{bio.var}{integer (vector). Input which monthly data should be downloaded. Only applicable to BIOCLIM variables. For further information see: \url{https://www.worldclim.org/data/bioclim.html}. \cr Default: \code{c(1:19)}}

\item{month.var}{integer (vector). Input which monthly data should be downloaded. \cr Default: \code{c(1:12)}}

\item{resolution}{string (vector). Ranging from a 10 arc-minute resolution over 5 and 2.5 arc-minute to 30 arc-second resolution.\cr Default: \code{c("10m", "5m", "2.5m", "30s")}}

\item{version.var}{string (vector). Input which version of the dataset should be downloaded. Multiple selection is possible. \cr Default:  \code{c("1.4", "2.1")}}

\item{clipping}{logical. Input whether the downloaded data should be clipped. See \code{\link{clipping.tif}} for more information. \cr If \code{FALSE}: \code{clip.shapefile}, \code{buffer}, \code{clip.extent} will be ignored. \cr Default: \code{FALSE}}

\item{clip.shapefile}{string. Input which shapefile should be used for clipping. \cr Default: \code{NULL}}

\item{clip.extent}{numeric (vector). Input vector with four numeric values. This is following the input order c("xleft", "xright", "ybottom", "ytop").\cr Default: \code{c(-180, 180, -90, 90)}}

\item{buffer}{numeric. Input of decimal degrees of buffer around the shapefile and/or extent. \cr Default: \code{0}}

\item{convert.files.to.asc}{logical. Input whether files should be converted into the ASCII format. See \code{\link{convert.to.asc}} for more information.\cr If \code{TRUE}: a new subdirectory is created and the rawdata is saved there. \cr If the parameter \code{clipping} is also \code{TRUE}: the clipped raster files are also saved as ASCII grids. \cr  Default: \code{FALSE}}

\item{stacking.data}{logical. Input whether the downloaded data should be stacked as a netCDF-rasterstack. See \code{\link{stacking.downloaded.data}} for more information. \cr Default: \code{FALSE}}

\item{keep.raw.zip}{logical. Should the downloaded raw-data be provided as "zip"-file. See \code{\link{combine.raw.in.zip}} for more information. \cr Default: \code{FALSE}}

\item{delete.raw.data}{logical. Should the downloaded raw-data be deleted. \cr If the "combine.raw.zip"-option is \code{TRUE}: raw-data is still available in the zipped file. \cr Default: \code{FALSE}}

\item{save.bib.file}{logical. Whether a BibTex-citation file of the dataset should be provided in the Working directory. See \code{\link{save.citation}} for more information. \cr Default: \code{TRUE}}
}
\value{
WorldClim climate datasets for the period of 1960-1990 (for v1.4) and/or 1970-2000 (for v2.1).
}
\description{
This function supports the download, pre-processing and management of the WorldClim historical climate dataset.\cr This comprises of monthly precipitation sums (mm), temperature (average, maximum, minimum; °C), monthly solar radiation sums (kJ m-2 day-1), wind speed (m s-1), water vapor pressure (kPa), and annual chracteristics (19 bioclimatic variables). Also an elevation raster is provided.\cr To allow pre-processing, clipping and buffering, conversion to ASCII-grids and stacking options are included.\cr Optional an output of a .bib-file of the cited literature can be retrieved.\cr For user convenience, saving directories will be created automatically. Also options to "zip" and/or delete the RAW-files are included.
}
\note{
Please note that  solar radiation, wind speed, water vapor pressure, bioclimatic parameters, and elevation raster are only provided by the WorldClim Version 2.1 (current version).

Please note also that the downloaded data for temperature and the therefore also the first eleven bioclim-variables are processed to °C with one significant decimal without offset and factor. Processing and conversion to other file-formats on a global dataset may take some time depending on the spatial resolution.
}
\examples{
\dontrun{
# Bioclim
WorldClim.HistClim.download(parameter = "bio",
                            bio.var = c(1,12),
                            resolution = "10min",
                            version.var = c("1.4", "2.1"))
# Precipitation
WorldClim.HistClim.download(parameter = "prec",
                             month.var = c(1,12),
                             resolution = "10min",
                             version.var = c("1.4", "2.1")
                             )
}

}
\references{
R. J. Hijmans, S. E. Cameron, J. L. Parra, et al. "Very high resolution interpolated climate surfaces for global land areas". In: _International Journal of Climatology_ 25.15 (2005), pp. 1965-1978. DOI: 10.1002/joc.1276. <URL: https://doi.org/10.1002/joc.1276>.

S. E. Fick and R. J. Hijmans. "WorldClim 2: new 1-km spatial resolution climate surfaces for global land areas". In: _International Journal of Climatology_ 37.12 (Okt. 2017), pp. 4302-4315. DOI: 10.1002/joc.5086. <URL:https://doi.org/10.1002/joc.5086>.
}
\author{
Helge Jentsch
}
